+++ v2/src/modules/address/ModuleAddressNftTransfers.vue
@@ -17,8 +17,8 @@
                     />
                 </v-row>
             </div>
-            <app-btn v-if="props.isOverview && !mdAndDown" text="More" isSmall icon="east" @click="goToNftTransfersPage"></app-btn>
-            <app-btn-icon v-if="props.isOverview && mdAndDown" icon="more_horiz" @click="goToNftTransfersPage"></app-btn-icon>
+            <app-btn v-if="props.isOverview && !xs" text="More" isSmall icon="east" @click="goToNftTransfersPage"></app-btn>
+            <app-btn-icon v-if="props.isOverview && xs" icon="east" @click="goToNftTransfersPage"></app-btn-icon>
         </v-card-title>
         <div>
             <!-- Table Header -->
@@ -30,7 +30,7 @@
                 <v-col v-if="!props.isOverview" cols="2" class="py-0"> Hash </v-col>
                 <v-col v-if="!props.isOverview" cols="2" class="py-0"> Timestamp </v-col>
             </v-row>
-            <v-divider class="my-0 mt-md-5 mx-n4 mx-sm-n6" />
+            <v-divider class="my-0 mt-sm-4 mx-n4 mx-sm-n6" />
             <div v-if="initialLoad || loadingMeta || loadingTransfers" class="p-ten-top">
                 <div v-for="item in 10" :key="item" style="padding: 10px 0">
                     <div class="skeleton-box rounded-xl" style="height: 40px"></div>
@@ -82,7 +82,7 @@ import { ITEMS_PER_PAGE } from '@core/constants'
 const OVERVIEW_MAX_ITEMS = 7
 const MOBILE_MAX_ITEMS = 4
 
-const { mdAndDown } = useDisplay()
+const { mdAndDown, xs } = useDisplay()
 const props = defineProps({
     addressHash: {
         type: String,