+++ v2/src/modules/tokens/ModuleTokensInfo.vue
@@ -1,10 +1,11 @@
 <template>
     <v-card variant="elevated" elevation="1" rounded="xl" class="pa-4 pa-sm-6 h-100">
-        <v-card-title :class="['pa-0', { 'd-flex align-center justify-space-between mb-4': isHomePage }]">
+        <v-card-title :class="['pa-0', { 'd-flex align-center justify-space-between mb-2 mb-sm-4': isHomePage }]">
             <h1 :class="[isHomePage ? 'text-h6' : 'text-h4 ', 'font-weight-bold']">
                 {{ title }}
             </h1>
-            <app-btn v-if="isHomePage" text="More" isSmall icon="east" @click="goToTokens"></app-btn>
+            <app-btn v-if="isHomePage && !xs" text="More" isSmall icon="east" @click="goToTokens"></app-btn>
+            <app-btn-icon v-else-if="isHomePage && xs" icon="east" @click="goToTokens"></app-btn-icon>
         </v-card-title>
         <app-tabs v-if="!isHomePage" v-model="state.activeList" :routes="routes" :tabs="list" class="my-5" btn-variant></app-tabs>
         <v-row v-if="!isHomePage" class="mb-10 flex-nowrap" align="center">
@@ -12,7 +13,7 @@
             <module-add-fav-token v-if="state.activeList === list[1].value" />
         </v-row>
         <v-row align="center" justify="start" class="text-body-1 text-info my-0 d-none d-sm-flex">
-            <v-col sm="6" :md="isHomePage ? '6' : '4'">
+            <v-col sm="6" :md="isHomePage ? '6' : '4'" class="py-0">
                 <v-row align="center" class="ma-0">
                     <v-btn variant="text" color="info" class="font-weight-regular ml-n3" rounded="pill" size="small" @click="sortTable(SORT_KEY.NAME)">
                         Token
@@ -20,7 +21,7 @@
                     >
                 </v-row>
             </v-col>
-            <v-col sm="2" md="2">
+            <v-col sm="2" md="2" class="py-0">
                 <v-row align="center" class="ma-0">
                     <v-btn variant="text" color="info" class="font-weight-regular ml-n3" rounded="pill" size="small" @click="sortTable(SORT_KEY.PRICE)">
                         Price
@@ -28,12 +29,12 @@
                     >
                 </v-row>
             </v-col>
-            <v-col sm="2" md="2">
+            <v-col sm="2" md="2" class="py-0">
                 <v-btn variant="text" color="info" class="font-weight-regular ml-n3" rounded="pill" size="small" @click="sortTable(SORT_KEY.PERCENTAGE_CHANGE)">
                     24h<v-icon v-if="isActiveSort(SORT_KEY.PERCENTAGE_CHANGE)" class="ml-1" :size="14">{{ sortIcon }}</v-icon></v-btn
                 ></v-col
             >
-            <v-col v-if="!isHomePage" md="2" class="d-none d-md-block">
+            <v-col v-if="!isHomePage" md="2" class="d-none d-md-block py-0">
                 <v-row align="center" class="ma-0">
                     <v-btn variant="text" color="info" class="font-weight-regular ml-n3" rounded="pill" size="small" @click="sortTable(SORT_KEY.VOLUME)">
                         Volume
@@ -41,7 +42,7 @@
                     >
                 </v-row>
             </v-col>
-            <v-col :sm="2" class="d-none d-sm-block">
+            <v-col :sm="2" class="d-none d-sm-block py-0">
                 <v-row align="center" class="ma-0">
                     <v-btn variant="text" color="info" class="font-weight-regular ml-n3" rounded="pill" size="small" @click="sortTable(SORT_KEY.MARKET_CAP)">
                         Market Cap
@@ -51,9 +52,9 @@
             </v-col>
             <!-- <v-col sm="2" lg="1"> Watchlist </v-col> -->
         </v-row>
-        <v-divider class="my-0 mt-md-4 mx-n4 mx-sm-n6" />
+        <v-divider class="my-0 mt-sm-3 mx-n4 mx-sm-n6" />
         <template v-if="!loadingCoinData">
-            <div v-if="tokensInPage.length > 0">
+            <div v-if="tokensInPage.length > 0" class="p-ten-top">
                 <div v-for="token in currentPageData" :key="token.contract">
                     <token-market-info-table-row v-if="token" :token="token" :is-home-view="isHomePage" />
                 </div>
@@ -76,6 +77,7 @@
 <script setup lang="ts">
 import AppTabs from '@core/components/AppTabs.vue'
 import AppBtn from '@core/components/AppBtn.vue'
+import AppBtnIcon from '@core/components/AppBtnIcon.vue'
 import AppInput from '@core/components/AppInput.vue'
 import AppNoResult from '@core/components/AppNoResult.vue'
 import AppPagination from '@core/components/AppPagination.vue'
@@ -93,6 +95,9 @@ import { searchHelper } from '@core/helper/search'
 import { useAppPaginate } from '@core/composables/AppPaginate/useAppPaginate.composable'
 import { TOKENS_VIEW } from './models/tokensView'
 import { useRouter } from 'vue-router'
+import { useDisplay } from 'vuetify'
+
+const { xs } = useDisplay()
 const routes = ADDRESS_ROUTE_QUERY.Q_NFTS
 
 // const MAX_TOKENS = 200